requirements.txt
fastapi[standard]==0.112.0
requests
python-dotenv
elasticsearch